% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DDEG.R
\name{run_DDEG}
\alias{run_DDEG}
\title{DDEG: Differential Data Expression Generator with WGCNA}
\usage{
run_DDEG(
  count_file,
  meta_file = NULL,
  group_column = NULL,
  p_val_cutoff = 0.05,
  logfc_cutoff = 1
)
}
\arguments{
\item{count_file}{Path to the count matrix (CSV or TXT)}

\item{meta_file}{(Optional) Path to the metadata file}

\item{group_column}{(Optional) The column name in metadata}

\item{p_val_cutoff}{Cutoff for adjusted p-value (default 0.05)}

\item{logfc_cutoff}{Cutoff for absolute log2 fold change (default 1)}
}
\description{
DDEG: Differential Data Expression Generator with WGCNA
}
